+++ src/initrd/mod.rs
@@ -4,6 +4,7 @@
 //! services are configured, so it may not be able to use all usual metadata
 //! fetcher.
 
+use crate::providers::kubevirt;
 use crate::providers::proxmoxve::ProxmoxVEConfigDrive;
 use crate::providers::vmware::VmwareProvider;
 use crate::providers::MetadataProvider;
@@ -19,6 +20,7 @@ pub(crate) fn fetch_network_kargs(provider: &str) -> Result<Option<String>> {
     match provider {
         "vmware" => VmwareProvider::try_new()?.rd_network_kargs(),
         "proxmoxve" => ProxmoxVEConfigDrive::try_new()?.rd_network_kargs(),
+        "kubevirt" => kubevirt::try_new_provider_else_noop()?.rd_network_kargs(),
         _ => Ok(None),
     }
 }