+++ src/metadata.rs
@@ -26,7 +26,7 @@ use crate::providers::gcp::GcpProvider;
 use crate::providers::hetzner::HetznerProvider;
 use crate::providers::ibmcloud::IBMGen2Provider;
 use crate::providers::ibmcloud_classic::IBMClassicProvider;
-use crate::providers::kubevirt::KubeVirtProvider;
+use crate::providers::kubevirt;
 use crate::providers::microsoft::azure::Azure;
 use crate::providers::microsoft::azurestack::AzureStack;
 use crate::providers::openstack;
@@ -68,7 +68,7 @@ pub fn fetch_metadata(provider: &str) -> Result<Box<dyn providers::MetadataProvi
         "ibmcloud" => box_result!(IBMGen2Provider::try_new()?),
         // IBM Cloud - Classic infrastructure.
         "ibmcloud-classic" => box_result!(IBMClassicProvider::try_new()?),
-        "kubevirt" => box_result!(KubeVirtProvider::try_new()?),
+        "kubevirt" => kubevirt::try_new_provider_else_noop(),
         "openstack" => openstack::try_config_drive_else_network(),
         "openstack-metadata" => box_result!(OpenstackProviderNetwork::try_new()?),
         "oraclecloud" => box_result!(OracleCloudProvider::try_new()?),